requirements.txt
robotframework >= 4.1.2
allure-robotframework >= 2.9.45